Modularize executable main functions into separate sources (#1160)
